argument-extractor/requirements.txt
accelerate<=0.28.0
fire<=0.6.0
flair<=0.13.1
jupyter<=1.0.0
openai<=1.14.0
pandas<=2.2.1
pydantic<=2.6.4
python-docx<=1.1.0
python-dotenv<=1.0.1
requests<=2.31.0
sentence-transformers<=2.5.1
tenacity<=8.2.3
tiktoken<=0.6.0
tqdm<=4.66.2
torch<=2.1.2
transformers<=4.38.2
wasabi<=1.1.2
wtpsplit<=1.2.4